refactor: reexport iroh_relay::dns module fully
