+++ iroh-relay/src/dns.rs
@@ -1,4 +1,10 @@
-//! DNS resolver
+//! Configurable DNS resolver for `iroh-relay` and `iroh`.
+//!
+//! The main export is the [`DnsResolver`] struct. It provides methods to resolve domain names
+//! to IPv4 and IPv6 addresses. Additionally, the resolver features methods to resolve the
+//! [`EndpointInfo`] for an iroh [`EndpointId`] from `_iroh` TXT records.
+//! See the [`endpoint_info`] module documentation for details on how iroh endpoint records
+//! are structured.
 
 use std::{
     fmt,
@@ -192,6 +198,11 @@ impl Builder {
 }
 
 /// The DNS resolver used throughout `iroh`.
+///
+/// By default, we use a built-in resolver that reads the system's DNS configuration.
+/// The nameservers can be customized by constructing the resolver with [`Self::builder`].
+/// Alternatively, you can create a fully custom DNS resolver by implementing the [`Resolver`]
+/// trait and creating the resolver with [`Self::custom`].
 #[derive(Debug, Clone)]
 pub struct DnsResolver(DnsResolverInner);
