+++ iroh/src/lib.rs
@@ -262,8 +262,6 @@ pub(crate) mod web_runtime;
 
 pub mod address_lookup;
 pub mod defaults;
-#[cfg(not(wasm_browser))]
-pub mod dns;
 pub mod endpoint;
 pub mod metrics;
 mod net_report;
@@ -274,6 +272,8 @@ pub use iroh_base::{
     EndpointAddr, EndpointId, KeyParsingError, PublicKey, RelayUrl, RelayUrlParseError, SecretKey,
     Signature, SignatureError, TransportAddr,
 };
+#[cfg(not(wasm_browser))]
+pub use iroh_relay::dns;
 pub use iroh_relay::{RelayConfig, RelayMap, endpoint_info};
 pub use n0_watcher::Watcher;
 pub use net_report::{Report as NetReport, TIMEOUT as NET_REPORT_TIMEOUT};
